# cmd/api/auth.go
package main

import (
	"net/http"
	"social/internal/store"
)

type RegisterUserPayload struct {
	Username string `json:"username" validate:"required,max=100"`
	Email    string `json:"email" validate:"required,email,max=255"`
	Password string `json:"password" validate:"required,min=3,max=72"`
}

type UserWithToken struct {
	*store.User
	Token string `json:"token"`
}

// registerUserHandler godoc
//
//	@Summary		Registers a user
//	@Description	Registers a user
//	@Tags			authentication
//	@Accept			json
//	@Produce		json
//	@Param			payload	body		RegisterUserPayload	true	"User credentials"
//	@Success		201		{object}	UserWithToken		"User registered"
//	@Failure		400		{object}	error
//	@Failure		500		{object}	error
//	@Router			/authentication/user [post]
func (app *application) registerUserHandler(w http.ResponseWriter, r *http.Request) {
	var payload RegisterUserPayload
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}

	// if err := Validate.Struct(payload); err != nil {
	// 	app.badRequestResponse(w, r, err)
	// 	return
	// }

	user := &store.User{
		Username: payload.Username,
		Email:    payload.Email,
	}

	// hash the user password
	if err := user.Password.Set(payload.Password); err != nil {
		app.internalServerError(w, r, err)
		return
	}

}

// type CreateUserTokenPayload struct {
// 	Email    string `json:"email" validate:"required,email,max=255"`
// 	Password string `json:"password" validate:"required,min=3,max=72"`
// }

// createTokenHandler godoc
//
//	@Summary		Creates a token
//	@Description	Creates a token for a user
//	@Tags			authentication
//	@Accept			json
//	@Produce		json
//	@Param			payload	body		CreateUserTokenPayload	true	"User credentials"
//	@Success		200		{string}	string					"Token"
//	@Failure		400		{object}	error
//	@Failure		401		{object}	error
//	@Failure		500		{object}	error
//	@Router			/authentication/token [post]
// func (app *application) createTokenHandler(w http.ResponseWriter, r *http.Request) {
// 	var payload CreateUserTokenPayload
// 	if err := readJSON(w, r, &payload); err != nil {
// 		app.badRequestResponse(w, r, err)
// 		return
// 	}

// 	if err := Validate.Struct(payload); err != nil {
// 		app.badRequestResponse(w, r, err)
// 		return
// 	}

// 	user, err := app.store.Users.GetByEmail(r.Context(), payload.Email)
// 	if err != nil {
// 		switch err {
// 		case store.ErrNotFound:
// 			app.unauthorizedErrorResponse(w, r, err)
// 		default:
// 			app.internalServerError(w, r, err)
// 		}
// 		return
// 	}

// 	if err := user.Password.Compare(payload.Password); err != nil {
// 		app.unauthorizedErrorResponse(w, r, err)
// 		return
// 	}

// 	claims := jwt.MapClaims{
// 		"sub": user.ID,
// 		"exp": time.Now().Add(app.config.auth.token.exp).Unix(),
// 		"iat": time.Now().Unix(),
// 		"nbf": time.Now().Unix(),
// 		"iss": app.config.auth.token.iss,
// 		"aud": app.config.auth.token.iss,
// 	}

// 	token, err := app.authenticator.GenerateToken(claims)
// 	if err != nil {
// 		app.internalServerError(w, r, err)
// 		return
// 	}

// 	if err := app.jsonResponse(w, http.StatusCreated, token); err != nil {
// 		app.internalServerError(w, r, err)
// 	}
// }